Delete old test runs (#10144)
* Delete old test runs

* Remove old tests; Rename to new test suite path

---------
