+++ packages/e2e-tests/helpers/authentication.ts
@@ -11,12 +11,6 @@ export const E2E_USER_1_API_KEY = "ruk_jukvxbSz7syp4Tw21RzEwSK2bjucNDklCEDmVkjbH
 // trunk-ignore(gitleaks/generic-api-key)
 export const E2E_USER_2_API_KEY = "ruk_K77rBZ4FuPyfj5ocfpPpTXJx7WGVesSySInnCJ4sS1Y";
 
-// Can view the "FE Test Golden Recording" (Test Suites) workspace
-// but cannot upload new recordings nor source-maps
-// trunk-ignore(gitleaks/generic-api-key)
-export const E2E_USER_3_API_KEY = "rwk_FnFcJoJ158eTLXJTg8gpXMto7WT3N8dhNKcAlo2cSHO";
-export const E2E_USER_3_TEAM_ID = "dzoxNzlkZGU4Mi00MmVhLTRkZTctYmI1OC04ZWY0NDQxZTcyYTg=";
-
 // Can view the "Replay Devtools Snapshots" (Test Suites) workspace
 // but cannot upload new recordings nor source-maps
 // trunk-ignore(gitleaks/generic-api-key)
