--- src/ui/components/Library/Team/View/NewTestRuns/Overview/TestRunOverviewContextRoot.tsx
@@ -1,10 +1,0 @@
-import { TestRunOverviewContainer } from "../../TestRuns/Overview/TestRunOverviewContainerContextType";
-import { TestRunOverviewContent } from "./TestRunOverviewContent";
-
-export function TestRunOverviewPage() {
-  return (
-    <TestRunOverviewContainer>
-      <TestRunOverviewContent />
-    </TestRunOverviewContainer>
-  );
-}
