--- src/ui/components/Library/Team/View/NewTestRuns/TestRunsPage.module.css
@@ -1,30 +1,0 @@
-.filterContainer {
-  position: relative;
-}
-
-.filterInput {
-  width: 100%;
-  padding: 0.25rem 0.625rem;
-  appearance: none;
-  border-radius: 0.5rem;
-  border: none;
-  background-color: var(--body-bgcolor);
-  border: 1px solid var(--input-border);
-  font-size: 0.75rem;
-  outline: none;
-  padding-right: 1.55rem;
-}
-
-.filterInput:focus {
-  box-shadow: 0 0 0 2px var(--primary-accent);
-}
-
-.searchIcon {
-  position: absolute;
-  right: 0.5rem;
-  top: 50%;
-  height: 1rem;
-  width: 1rem;
-  transform: translateY(-50%);
-  opacity: 0.5;
-}
