+++ src/ui/components/Library/Team/View/TestRuns/Overview/RunResults.tsx
@@ -5,71 +5,35 @@ import {
   memo,
   useContext,
   useDeferredValue,
+  useMemo,
   useState,
 } from "react";
 
 import Icon from "replay-next/components/Icon";
-import { useTestRunDetailsSuspends } from "ui/components/Library/Team/View/TestRuns/hooks/useTestRunDetailsSuspends";
 import {
   FileNode,
   PathNode,
   isPathNode,
+  treeContainTest,
   useFileNameTree,
-} from "ui/components/Library/Team/View/TestRuns/Overview/useFileNameTree";
+} from "ui/components/Library/Team/View/TestRuns/hooks/useFileNameTree";
+import { useTestRunDetailsSuspends } from "ui/components/Library/Team/View/TestRuns/hooks/useTestRunDetailsSuspends";
 import { TestRunsContext } from "ui/components/Library/Team/View/TestRuns/TestRunsContextRoot";
-import { TestGroup, TestGroups } from "ui/utils/testRuns";
+import FileIcon from "ui/components/shared/Icon";
+import { TestGroup } from "ui/utils/testRuns";
 
-import { TestResultListItem } from "./TestResultListItem";
-import styles from "../../../../Testsuites.module.css";
+import styles from "../TestRuns.module.css";
 
-function hasDuplicateRecordings(fileNameToTests: TestGroup["fileNameToTests"]) {
-  return Object.values(fileNameToTests).some(tests => {
-    const recordingIds = new Set<string>();
-    for (const test of tests) {
-      for (const execution of test.executions) {
-        for (const recording of execution.recordings) {
-          if (recordingIds.has(recording.id)) {
-            return true;
-          } else {
-            recordingIds.add(recording.id);
-          }
-        }
-      }
-    }
-
-    return false;
-  });
-}
-
-function DeepLinkWarning({ testGroups }: { testGroups: TestGroups }) {
-  const duplicate =
-    hasDuplicateRecordings(testGroups.failedRecordings.fileNameToTests) ||
-    hasDuplicateRecordings(testGroups.passedRecordings.fileNameToTests) ||
-    hasDuplicateRecordings(testGroups.flakyRecordings.fileNameToTests);
-
-  if (!duplicate) {
-    return null;
-  }
-
-  return (
-    <div
-      className="m-2 flex flex-row items-center gap-2 rounded-lg border p-2"
-      style={{
-        color: "var(--theme-warning-color)",
-        backgroundColor: "var(--theme-warning-background)",
-        borderColor: "var(--theme-warning-border)",
-      }}
-    >
-      <Icon type="warning" /> Heads up! Deep linking to tests within a recording will be added soon.
-    </div>
-  );
-}
-
-export function RunResults({ isPending }: { isPending: boolean }) {
+export function RunResults({
+  testFilterByText,
+  filterCurrentRunByStatus,
+}: {
+  testFilterByText: string;
+  filterCurrentRunByStatus: "all" | "failed-and-flaky";
+}) {
   const { testRunId } = useContext(TestRunsContext);
 
-  const [filterByText, setFilterByText] = useState("");
-  const filterByTextDeferred = useDeferredValue(filterByText);
+  const filterByTextDeferred = useDeferredValue(testFilterByText);
 
   const { groupedTests } = useTestRunDetailsSuspends(testRunId);
   assert(groupedTests !== null);
@@ -77,49 +41,29 @@ export function RunResults({ isPending }: { isPending: boolean }) {
   const { passedRecordings, failedRecordings, flakyRecordings } = groupedTests;
 
   return (
-    <>
-      <div
-        className={`relative mb-2 border-b border-themeBorder bg-bodyBgcolor p-2 ${
-          isPending ? "opacity-50" : ""
-        }`}
-      >
-        <input
-          className={`w-full appearance-none rounded border-none text-xs focus:outline-none focus:ring focus:ring-primaryAccent ${styles.FilterTestsFilter}`}
-          data-test-id="TestRunResults-FilterInput"
-          onChange={event => setFilterByText(event.currentTarget.value)}
-          placeholder="Filter tests"
-          type="text"
-          value={filterByText}
-        />
-
-        <Icon
-          className="absolute right-4 top-1/2 h-4 w-4 -translate-y-1/2 opacity-50"
-          type="search"
-        />
-      </div>
-      <div
-        className={`flex flex-col overflow-y-auto ${isPending ? "opacity-50" : ""}`}
-        data-filtered-by-text={filterByTextDeferred}
-        data-test-id="TestRunResults"
-      >
-        <DeepLinkWarning testGroups={groupedTests} />
-        <TestStatusGroup
-          filterByText={filterByTextDeferred}
-          label="Failed"
-          testGroup={failedRecordings}
-        />
-        <TestStatusGroup
-          filterByText={filterByTextDeferred}
-          label="Flaky"
-          testGroup={flakyRecordings}
-        />
+    <div
+      className="flex flex-col overflow-y-auto"
+      data-filtered-by-text={filterByTextDeferred}
+      data-test-id="TestRunResults"
+    >
+      <TestStatusGroup
+        filterByText={filterByTextDeferred}
+        label="Failed"
+        testGroup={failedRecordings}
+      />
+      <TestStatusGroup
+        filterByText={filterByTextDeferred}
+        label="Flaky"
+        testGroup={flakyRecordings}
+      />
+      {filterCurrentRunByStatus !== "failed-and-flaky" && (
         <TestStatusGroup
           filterByText={filterByTextDeferred}
           label="Passed"
           testGroup={passedRecordings}
         />
-      </div>
-    </>
+      )}
+    </div>
   );
 }
 
@@ -147,15 +91,18 @@ function TestStatusGroup({
       data-test-id={`TestRunResults-StatusGroup-${label.toLowerCase()}`}
     >
       <div
+        data-test-id="TestRunResults-StatusGroup-Title"
         className={`top-0 flex grow flex-row p-2 pl-4 font-medium hover:cursor-pointer ${styles.libraryRowHeader}`}
         onClick={() => setExpanded(!expanded)}
       >
-        <div className="grow">
+        <div data-status={label} className={`grow font-bold ${styles.testStatusHeader}`}>
           <span data-test-id="TestRunResults-StatusGroup-Count">{count}</span> {label} Test
           {count > 1 ? "s" : ""}
         </div>
         <div className="flex">
           <Icon
+            data-test-id="TestRunResults-StatusGroup-Icon"
+            data-test-state={expanded ? "expanded" : "collapsed"}
             className={`${
               expanded ? "" : "rotate-90"
             } rotate duration-140 h-4 w-4 transition ease-out`}
@@ -172,57 +119,49 @@ function TestStatusGroup({
 
 const FileNodeRenderer = memo(function FileNodeRenderer({
   depth,
-  filterByText,
   label,
   fileNode,
 }: {
   depth: number;
-  filterByText: string;
   label: string;
   fileNode: FileNode;
 }) {
-  const { name, test, nestedRecordingCount } = fileNode;
+  const { test } = fileNode;
+  const currentTestId = test.testId;
+  const { setTestId, testId } = useContext(TestRunsContext);
 
-  const [expanded, setExpanded] = useState(true);
+  const onClick = () => setTestId(currentTestId);
 
-  const onClick = () => setExpanded(!expanded);
+  let iconFilename: string;
+  let iconClass: string;
+  if (label === "Passed") {
+    iconFilename = "testsuites-success";
+    iconClass = styles.testsuitesSuccess;
+  } else if (label === "Failed") {
+    iconClass = styles.testsuitesFailed;
+    iconFilename = "testsuites-v2-failed";
+  } else {
+    iconClass = styles.testsuitesFlaky;
+    iconFilename = "testsuites-v2-flaky";
+  }
+
+  const isSelected = currentTestId === testId;
 
   return (
     <>
       <div
-        className={`flex cursor-pointer items-center gap-2 truncate  py-2  pr-4 ${styles.libraryRow}`}
+        className={`flex cursor-pointer items-center gap-2 truncate rounded py-1.5 pr-4 ${
+          styles.libraryRow
+        } ${isSelected ? styles.libraryRowSelected : ""}`}
         data-test-id="TestRunResult-FileNode"
-        data-test-state={expanded ? "expanded" : "collapsed"}
         onClick={onClick}
         style={{
           paddingLeft: `${depth * 1}rem`,
         }}
       >
-        <Icon className="h-5 w-5 shrink-0" type="file" />
-        <div className="truncate">{name}</div>
-        {!expanded && (
-          <div className="text-xs text-bodySubColor">
-            ({nestedRecordingCount} {nestedRecordingCount === 1 ? "test" : "tests"})
-          </div>
-        )}
+        <FileIcon className={iconClass} filename={iconFilename} />
+        <div className="truncate">{fileNode.name}</div>
       </div>
-      <Offscreen mode={expanded ? "visible" : "hidden"}>
-        {test.executions
-          .filter(e => e.recordings.length > 0)
-          .flatMap(execution =>
-            execution.recordings.map(recording => (
-              <TestResultListItem
-                depth={depth + 1}
-                filterByText={filterByText}
-                key={test.id + recording.id}
-                label={execution.result}
-                recording={recording}
-                test={test}
-                secondaryBadgeCount={/* index > 0 ? index + 1 : null */ null}
-              />
-            ))
-          )}
-      </Offscreen>
     </>
   );
 });
@@ -239,8 +178,16 @@ function PathNodeRenderer({
   pathNode: PathNode;
 }) {
   const { children, name, pathNames } = pathNode;
-
   const [expanded, setExpanded] = useState(true);
+  const { testId } = useContext(TestRunsContext);
+
+  const containsSelectedSpec = useMemo(() => {
+    if (expanded || !testId) {
+      return false;
+    }
+
+    return treeContainTest(children, testId);
+  }, [children, expanded, testId]);
 
   const onClick = () => setExpanded(!expanded);
 
@@ -261,20 +208,29 @@ function PathNodeRenderer({
     <>
       {name && (
         <div
-          className={`flex cursor-pointer items-center gap-2 truncate py-2 pr-4 ${styles.libraryRow}`}
+          className={`cursor-pointer truncate rounded py-2 pr-4 ${styles.libraryRow} ${
+            containsSelectedSpec ? styles.libraryRowSelected : ""
+          }`}
           data-test-id="TestRunResult-PathNode"
           data-test-state={expanded ? "expanded" : "collapsed"}
           onClick={onClick}
           style={{
-            paddingLeft: `${depth * 0.5}rem`,
+            paddingLeft: `${depth * 1}rem`,
           }}
         >
-          <Icon className="h-5 w-5 shrink-0" type={expanded ? "folder-open" : "folder-closed"} />
-          <div className="flex items-center gap-1 truncate">{formattedNames}</div>
+          <div className="flex items-center gap-2">
+            <div className="flex items-center gap-1 truncate">{formattedNames}/</div>
+            <Icon
+              data-test-id="TestRunResult-PathNode-Icon"
+              data-test-state={expanded ? "expanded" : "collapsed"}
+              className={`${
+                expanded ? "" : "rotate-90"
+              } rotate duration-140 h-4 w-4 transition ease-out`}
+              type="chevron-down"
+            />
+          </div>
           {!expanded && (
-            <div className="text-xs text-bodySubColor">
-              ({pathNode.nestedTestCount} {pathNode.nestedTestCount === 1 ? "test" : "tests"})
-            </div>
+            <div className="text-xs text-bodySubColor">({pathNode.nestedTestCount} tests)</div>
           )}
         </div>
       )}
@@ -292,13 +248,7 @@ function PathNodeRenderer({
             );
           } else {
             return (
-              <FileNodeRenderer
-                depth={depth + 1}
-                filterByText={filterByText}
-                key={index}
-                label={label}
-                fileNode={childNode}
-              />
+              <FileNodeRenderer depth={depth + 1} key={index} label={label} fileNode={childNode} />
             );
           }
         })}
