+++ src/ui/components/Library/Team/View/TestRuns/Overview/TestResultListItem.tsx
@@ -1,36 +1,69 @@
+import { motion } from "framer-motion";
 import { useRouter } from "next/router";
 
 import Icon from "replay-next/components/Icon";
 import { Recording } from "shared/graphql/types";
-import { TestRunTest } from "shared/test-suites/TestRun";
-import HighlightedText from "ui/components/Library/Team/View/TestRuns/HighlightedText";
+import { TestRun, TestRunTest } from "shared/test-suites/TestRun";
+import { trackEvent } from "ui/utils/telemetry";
 
+import {
+  getDurationString,
+  getTruncatedRelativeDate,
+} from "../../Recordings/RecordingListItem/RecordingListItem";
 import { StatusIcon } from "../../Tests/Overview/StatusIcon";
+import { AttributeContainer } from "../AttributeContainer";
 import styles from "../../../../Testsuites.module.css";
+
+function RecordingAttributes({
+  recording,
+  testRun,
+}: {
+  recording: Recording;
+  testRun: TestRun | null;
+}) {
+  return (
+    <div className="flex flex-row flex-wrap items-center gap-4 text-xs">
+      <AttributeContainer
+        dataTestId="Recording-Date"
+        icon="schedule"
+        title={new Date(recording.date).toLocaleString()}
+      >
+        {getTruncatedRelativeDate(recording.date)}
+      </AttributeContainer>
+
+      {recording.duration && recording.duration > 0 ? (
+        <AttributeContainer dataTestId="Recording-Duration" icon="timer">
+          {getDurationString(recording.duration)}
+        </AttributeContainer>
+      ) : null}
+      {testRun?.source?.user ? (
+        <AttributeContainer dataTestId="Recording-Username" icon="person">
+          {testRun.source.user}
+        </AttributeContainer>
+      ) : null}
+    </div>
+  );
+}
 
 export function TestResultListItem({
   depth,
-  filterByText,
   label,
   recording,
-  secondaryBadgeCount,
+  testRun,
   test,
 }: {
   depth: number;
-  filterByText: string;
-  label: string;
+  label: TestRunTest["result"];
   recording: Recording;
-  secondaryBadgeCount: number | null;
+  testRun: TestRun | null;
   test: TestRunTest;
 }) {
   const { comments, isProcessed, id: recordingId } = recording;
-  const { sourcePath, title } = test;
+  const { title } = test;
 
   const { apiKey, e2e } = useRouter().query;
 
   const numComments = comments?.length ?? 0;
-
-  label = label.toLowerCase();
 
   return (
     <a
@@ -38,21 +71,12 @@ export function TestResultListItem({
       className={`${styles.recordingLink} ${styles.libraryRow}`}
       data-test-id="TestRunResultsListItem"
       data-test-status={label}
-      style={{
-        paddingLeft: `${depth * 1}rem`,
-      }}
+      onClick={() => trackEvent("test_dashboard.open_replay", { view: "runs", result: label })}
     >
-      <div className={styles.linkContent}>
-        {secondaryBadgeCount != null && <Icon className={`${styles.icon}`} type="arrow-nested" />}
-        <StatusIcon isProcessed={isProcessed} status={label} />
-      </div>
-      <div className={styles.fileInfo}>
+      <StatusIcon status={label} isProcessed={isProcessed} />
+      <div className={`${styles.fileInfo} gap-1`}>
         <div className={styles.title}>{title || "Test"}</div>
-        {sourcePath && (
-          <div className={styles.filePath}>
-            <HighlightedText haystack={sourcePath} needle={filterByText} />
-          </div>
-        )}
+        <RecordingAttributes recording={recording} testRun={testRun} />
       </div>
       {numComments > 0 && (
         <div className={styles.comments}>
