+++ src/ui/components/Library/Team/View/TestRuns/Overview/TestRunOverviewContextRoot.tsx
@@ -1,4 +1,4 @@
-import { TestRunOverviewContainer } from "./TestRunOverviewContainerContextType";
+import { TestRunOverviewContainer } from "../TestRunOverviewContainerContextType";
 import { TestRunOverviewContent } from "./TestRunOverviewContent";
 
 export function TestRunOverviewPage() {
