+++ src/ui/components/Library/Team/View/TestRuns/RunStats.tsx
@@ -1,6 +1,6 @@
 import { useTestRunDetailsSuspends } from "ui/components/Library/Team/View/TestRuns/hooks/useTestRunDetailsSuspends";
 
-import styles from "../../../Testsuites.module.css";
+import styles from "./TestRuns.module.css";
 
 export function RunStats({ testRunId }: { testRunId: string }) {
   const { groupedTests } = useTestRunDetailsSuspends(testRunId);
@@ -16,18 +16,22 @@ export function RunStats({ testRunId }: { testRunId: string }) {
 
   return (
     <div className="flex shrink space-x-2">
-      {failed > 0 && <Pill className={styles.failedPill} value={failed} />}
-      {flakyCount > 0 && <Pill className={styles.flakyPill} value={flakyCount} />}
-      {passed > 0 && <Pill className={styles.successPill} value={passed} />}
+      {failed > 0 && <Pill status="failed" value={failed} />}
+      {flakyCount > 0 && <Pill status="flaky" value={flakyCount} />}
+      {passed > 0 && <Pill status="success" value={passed} />}
     </div>
   );
 }
 
-function Pill({ className, value }: { className: string; value: number }) {
+const pillStyles = {
+  success: styles.successPill,
+  failed: styles.failedPill,
+  flaky: styles.flakyPill,
+};
+
+function Pill({ status, value }: { status: "failed" | "success" | "flaky"; value: number }) {
   return (
-    <div
-      className={`flex h-[1.35rem] min-w-[1.35rem] items-center justify-center rounded-md text-xs font-bold ${className}`}
-    >
+    <div data-test-id={`Pill-${status}`} className={`${styles.Pill} ${pillStyles[status]}`}>
       {value}
     </div>
   );
