+++ src/ui/components/Library/Team/View/TestRuns/TestRunListItem.tsx
@@ -1,15 +1,13 @@
 import { useContext } from "react";
 
 import { TestRun, getTestRunTitle } from "shared/test-suites/TestRun";
-import { BranchIcon } from "ui/components/Library/Team/View/TestRuns/BranchIcon";
 import HighlightedText from "ui/components/Library/Team/View/TestRuns/HighlightedText";
 import Icon from "ui/components/shared/Icon";
+import MaterialIcon from "ui/components/shared/MaterialIcon";
 
 import { getTruncatedRelativeDate } from "../Recordings/RecordingListItem/RecordingListItem";
-import { AttributeContainer } from "./AttributeContainer";
-import { ModeAttribute } from "./Overview/RunSummary";
 import { TestRunsContext } from "./TestRunsContextRoot";
-import styles from "../../../Testsuites.module.css";
+import styles from "./TestRuns.module.css";
 
 function Status({ failCount }: { failCount: number }) {
   const status = failCount > 0 ? "fail" : "success";
@@ -17,7 +15,7 @@ function Status({ failCount }: { failCount: number }) {
   if (failCount > 0) {
     return (
       <div
-        className="flex h-5 w-5 shrink-0 items-center justify-center rounded-md bg-[#F02D5E] text-xs font-bold text-chrome"
+        className="flex h-5 w-8 shrink-0 items-center justify-center rounded-md bg-[color:var(--testsuites-v2-failed-pill)] text-xs font-bold text-white text-chrome"
         data-test-status={status}
       >
         {failCount}
@@ -25,7 +23,7 @@ function Status({ failCount }: { failCount: number }) {
     );
   } else {
     return (
-      <div data-test-status={status}>
+      <div className="flex w-8 flex-shrink-0 flex-col items-center" data-test-status={status}>
         <Icon className={styles.testsuitesSuccess} filename={"testsuites-success"} size="medium" />
       </div>
     );
@@ -39,7 +37,7 @@ export function TestRunListItem({
   filterByText: string;
   testRun: TestRun;
 }) {
-  const { date, source } = testRun;
+  const { date } = testRun;
 
   const { selectTestRun, testRunIdForDisplay } = useContext(TestRunsContext);
 
@@ -48,38 +46,6 @@ export function TestRunListItem({
   const failCount = testRun.results.counts.failed;
   const isSelected = testRunIdForDisplay === testRun.id;
 
-  let attributes;
-  if (source) {
-    const { branchName, isPrimaryBranch, user } = source;
-
-    attributes = (
-      <div className="flex flex-row items-center gap-4 text-xs font-light">
-        <AttributeContainer dataTestId="TestRun-Date" icon="schedule" title={date.toLocaleString()}>
-          {getTruncatedRelativeDate(date)}
-        </AttributeContainer>
-        {user && (
-          <AttributeContainer dataTestId="TestRun-Username" icon="person">
-            <HighlightedText haystack={user} needle={filterByText} />
-          </AttributeContainer>
-        )}
-        <BranchIcon
-          branchName={<HighlightedText haystack={branchName || ""} needle={filterByText} />}
-          isPrimaryBranch={isPrimaryBranch ?? false}
-          title={title}
-        />
-        <ModeAttribute testRun={testRun} />
-      </div>
-    );
-  } else {
-    attributes = (
-      <div className="flex flex-row items-center gap-4 text-xs font-light">
-        <AttributeContainer dataTestId="TestRun-Date" icon="schedule">
-          {getTruncatedRelativeDate(date)}
-        </AttributeContainer>
-      </div>
-    );
-  }
-
   const onClick = () => {
     selectTestRun(testRun.id);
   };
@@ -87,7 +53,7 @@ export function TestRunListItem({
   return (
     <div
       data-test-id="TestRunListItem"
-      className={`flex cursor-pointer flex-row items-center space-x-3 rounded-sm border-b border-chrome p-3 ${
+      className={`flex cursor-pointer flex-row items-center space-x-3 rounded-md bg-themeBase-100 px-2 py-1 ${
         styles.libraryRow
       }
       ${isSelected ? styles.libraryRowSelected : ""}
@@ -95,17 +61,20 @@ export function TestRunListItem({
       onClick={onClick}
     >
       <Status failCount={failCount} />
-
-      <div className="flex h-full flex-grow flex-col justify-evenly overflow-hidden">
-        <div className="flex flex-row justify-between space-x-3">
-          <div
-            className="wrap flex shrink grow-0 overflow-hidden text-ellipsis whitespace-nowrap pr-2 font-medium"
-            data-test-id="TestRun-Title"
-          >
-            <HighlightedText haystack={title} needle={filterByText} />
-          </div>
+      <div className="flex h-full flex-grow flex-row justify-between overflow-hidden">
+        <div className="wrap flex shrink grow-0 truncate pr-2" data-test-id="TestRun-Title">
+          <HighlightedText haystack={title} needle={filterByText} />
         </div>
-        {attributes}
+        <div
+          className="flex flex-shrink-0 items-center space-x-0.5 overflow-hidden text-ellipsis"
+          data-test-id="TestRun-Date"
+          title={title ?? ""}
+        >
+          <MaterialIcon className="w-4">schedule</MaterialIcon>
+          <span className="block overflow-hidden text-ellipsis whitespace-pre text-xs font-light">
+            {getTruncatedRelativeDate(date, false)}
+          </span>
+        </div>
       </div>
     </div>
   );
