+++ src/ui/components/Library/Team/View/TestRuns/TestRunsPage.module.css
@@ -1,28 +1,5 @@
-.dropdownTrigger {
-  display: flex;
-  cursor: pointer;
-  flex-direction: row;
-  align-items: center;
-  gap: 0.5rem;
-  border-radius: 0.375rem;
-  border: 1px solid var(--input-border);
-  background-color: var(--theme-text-field-bgcolor);
-  padding: 0.375rem 0.625rem;
-  font-size: 0.75rem;
-  color: var(--theme-text-field-color);
-  outline: none;
-  --ring-width: 3px;
-  --ring-color: var(--primary-accent);
-  white-space: nowrap;
-}
-
-.dropdownTrigger:focus {
-  box-shadow: 0 0 0 var(--ring-width) var(--ring-color);
-}
-
 .filterContainer {
   position: relative;
-  flex-grow: 1;
 }
 
 .filterInput {
